requirements.txt
langchain
streamlit
openai
elevenlabs
chromadb
google-search-results
google-api-python-client
tiktoken
python-decouple